Code refactor of existing search with no change in functionality. Removed unused Mitigation endpoints and associated services.
