+++ app/domain/__init__.py
@@ -1,0 +1,2 @@
+from  app.domain.PSQLTxt import PSQLTxt
+from app.domain.ParsedSearchString import ParsedSearchString
